fix: type mismatch

+++ src/model/Order.java
@@ -226,7 +226,7 @@ public class Order
 
 	public static double getTotalByOrderId(int orderId)
 	{
-		int orderTotalPrice = 0;
+		double orderTotalPrice = 0;
 		String query = "SELECT * FROM orderitems JOIN menuitems ON orderitems.menuItemId = menuitems.menuItemId WHERE orderitems.orderId = ?;";
 		  
 		try (Connection connection = Connect.getInstance().getConnection())
